requirements.txt
numpy~=1.19.5
opencv-python~=4.5.1.48
imutils~=0.5.4
screeninfo~=0.6.7
pygame~=2.0.1
matplotlib~=3.3.3
pyrebase4~=4.4.3
argparse~=1.4.0